Project/requirements.txt
psycopg2_binary==2.9.3
pandas==1.5.3
sqlalchemy==1.4.39
pyarrow==11.0.0
numpy==1.22.2
matplotlib==3.6.0
seaborn==0.12.2
scikit_learn==1.2.2
